Embeds $ORIGIN rpath into QtCore and libICU libraries
In case if Qt is configured without "-R ." which adds the $ORIGIN
rpath value, the linker will not try to find the copied over ICU
libraries (because the RPATH of a library higher in the dependency
tree does not get inherited by its children).

We want to make sure that the linker does attempt to find the ICU
libraries in the destination Qt libdir, so we prepend the
additional rpath value to the QtCore library, and also to the
ICU libraries.

Change-Id: Idbbf578d58ee12806b61610e6fd21f7c1ac48e3d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>