PEP 384-8-HT-2: Modify all type declarations
Status on 2018-05-04 macOS subset build:
Python 2: all works.
Python 3 release (PEP): all works.
Python 3 debug: untested.

In this second step, we will not introduce the heap types but do a preliminary
task: All the "&typename" entries will be renamed, and new variables
"typenameP" will  be introduced that point to these renamed structures.
The result of this action should be totally equivalent to the current state.

By this decoupling, we can later easily insert calls to PyType_Type.tp_new
instead of the references to the renamed entries and fix the resulting
bugs.

Task-number: PYSIDE-595
Change-Id: I81145ee891a5238c23110c598a8e96104ab2d247