Fix usage of PyCFunction_GET_FLAGS in pysidesignal.cpp
Using PyCFunction_GET_FLAGS is unsafe, because it does not check
whether the argument given is actually a PyCFunction object. This
macro needs to be replaced with the function equivalent
PyCFunction_GetFlags for stable ABI patch, but this will cause a crash
in PyCFunction_Call because we don't check the return value of
PyCFunction_GetFlags to see if it fails.

Rather than checking the return value, it is safe to preemptively
add a PyCFunction_Check before calling the GetFlags function. This
does not modify the logic behind signalCall function.

The crashing test was homonymoussignalandmethod_test.py, so no
new test is needed.

Task-number: PYSIDE-593
Change-Id: Id9ac9c0dec454e8e1ce9516dc68af924372a34a9
Reviewed-by: Christian Tismer <tismer@stackless.com>