PEP 384-8-HT-3: First successful heaptype transform
Status on 2018-05-04 macOS subset build:
Python 2: untested.
Python 3 release (PEP): all works.
Python 3 debug: untested.

In this third step, we do some serious transformation into
heaptypes. When this runs stable, we will generalize the solution
and convert the rest.

pysidesignal.cpp has problems with the heaptypes.
It is not clear yet what happens there.

PySideSignalMetaType already works fine, now also in debug builds.

Update:
The current approach seems not to work, because the creation of
heaptypes has certain assumptions which are hard to circumvent.
Without breaking the ABI, this seems to be impossible.

Task-number: PYSIDE-595
Change-Id: I06429cfb88c7037cc1cb8c5217ac2d1c438dc453