Fix wheel packages on macOS to contain min correct deployment target
Python's distutils does not know about the minimum deployment target
of extensions, but only the version that was used when building
the Python interpreter itself. This value is propagated to the final
wheel name, and confuses users on which macOS version can the package
be used.

This patch computes the correct minimum deployment target, and forces
distutils (and wheel) to use it in the wheel name. Unfortunately it
requires a slightly dirty hack to circumvent a rigid check in
distutils itself.

Task-number: PYSIDE-612
Change-Id: I4380ee81840f9746b210579700cc5d69d3865810
Reviewed-by: Fredrik Averpil <fredrik@averpil.com>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>