setup.py: Distinguish debug/release Python
When developing the heap types, I had to switch between debug and
release Python very often.

This patch allows for debug and release to co-exist without re-building
everything.

Additionally to the 'd' that gets appended to the build and install folders,
I used a 'p' for debug Python. Please feel free to change that as you like.

Change-Id: I022face5177bb69589809c8b235f09161d7fabd2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>