Improve QtInfo class
Previously the class would do a qmake process invocation for each
qmake -query property, after this change only a single invocation is
done to get all the query properties.

In addition another new invocation is done to find all mkspecs
variable, in order to find what kind of Qt build is being used:
a debug build, release build, or debug_and_release build.

This information is useful for packaging purposes, to know which files
should be copied or ignored.

Change-Id: If1ee4d19e0cc550e40dc568d1256030c8928c4e5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>