PEP 384-8: Enable the Limited Windows Build
Status on 2018-05-03 macOS subset build:
Python 2: all works.
Python 3 release (PEP): all works.
Python 3 debug: no tested.

This series of checkins is now ready for review.
Friedemann has found the last incompatibility on Windows (many thanks!),
and now it should finally be working!

Missing:
- object.h needs the heap type patch,

The signature module had some spots were it was quite incompatible with
PEP384. There could be almost all removed by using the fact that this is
Python 3 and we therefore have the __qualname__ attribute.

In the end, the only remaining wart was the usage of tp_dict, which I
replaced by a call to getattr("__dict__"), which gives us only a dictproxy.
I cheated and used the unpublished type 'mappingproxyobject'.

Instead of forcing a manual test of the validity of this breakage of
PEP384, I added code that computes the position of tp_dict and
tp_methods and ensures that they are exactly 4 pointers apart.

I hope that this is sufficient as a validation.

Furthermore unicodeobject.h was solved which was a bit tricky.

Solved:
    unicodeobject.h
    longobject.h
    descrobject.h
    pystate.h
    funcobject.h
    classobject.h
    abstract.h
    pythonrun.h
    code.h
    methodobject.h
    memoryobject.h
    bytesobject.h
    datetime.h
    pydebug.h
    floatobject.h
    tupleobject.h
    listobject.h

Unsolved:
    object.h

Task-number: PYSIDE-560
Change-Id: I5e8458eaf6280f5cde1dc46f58f78e8d79c7f08b