Fail early in setup.py when clang can not be found
Previously only a warning was printed when clang could not be
found at setup.py time, resulting in a not very nice to parse
CMake error.

Make sure to fail early in setup.py if clang can not be found.

Also make sure to check that the clang source variable0 is not
"None", because passing "None" to run_process_output results
in an even more obscure Python error.

Change-Id: Ia94bf7da51996a3d9c74d5d9978b1bf9e26b03d5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>