Upgrade shiboken diagrams and documentation
Applied Qt-style to the diagrams: shapes, and colors.
Removed references to Boost.
Move the .rst links to the bottom of the files for better
maintainability.
Removed unused images and wrapped a few long lines.
Aspell on the rst files.

The links needs to be updated once we have the URL for
the official documentation, since there are references
to external documentation source directories.

Task-number: PYSIDE-622
Change-Id: I2d553a29693ca4457e73a4ce5d78a2150501841d
Reviewed-by: Christian Tismer <tismer@stackless.com>