Update scriptableapplication example
Absolute paths are used to link the PySide2 libraries on Linux,
since it's not possible to ship symbolic links inside a wheel.
The README.txt was renamed to README.md to allow
syntax highlight on modern editors and also to be
compatible with online platforms.
The README.CMake.txt was merge to the README.md to include the
instruction on the same file.

Change-Id: Ie0fcb8cda770ff552576f6014b5822f8d278bfe6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>