QtXmlToSphinx: Refactor link handling
Forward declare QtXmlToSphinx::LinkContext as a public struct
and move the definition to the source file.

Replace the linkTag/linkTagEnding strings by an enumeration for the
type and flags for storing the bold/italic state.

Move the formatting to an operator<<(QTextStream,LinkContext)
for clarity.

Streamline the code checking the link text against the reference
in handleLinkText(), avoiding some unnecessary tests.

Add an 'External' type for external pages which do not have a role
keyword.

This fixes:
- Undoes a breakage introduced by c5582d3120249a902681454b2319df765ffa2cd8
  converting all link refs to RST labels, clobbering the '~' used for some
  function references. This is now limited to page type links.
- Broken handling of links within italic/bold: The assignment
  to the linkTag variable by type would clobber the previously assigned
  italic/bold markers
- Enable external links without role

Task-number: PYSIDE-363
Change-Id: I2e9a8cd80df93a77f3c6c9f8593243b35eeeb56e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>