PEP 384-9: Finish Limited Access to Type Objects
As a last step, the limited API has to be activated in PySide.

That sounded simple at first sight, but:
The function PyType_GetSlot works only on heap types.
But all our types inherit from existing types and need access
to these structures, too.

Instead, a type structure with all the needed fields was created.
It serves as a surrogate for the opaque type definition.
A validation routine proves at runtime if the assumptions
about the type structure are still plausible.

This is now working in all configurations.
Missing:

- cleanup,
- replace Sbk...Type by PyTypeObject,
- verification of the type heuristic.

Task-number: PYSIDE-560
Change-Id: I244fe1e6ed1a4cacbec7505a5572d6bfc4b86c6b