PEP 384-10: Remove the PyHeapType objects from shiboken
This set has the latest changes from the squash integrated
so everything is in sync, again.

The structures which used PyHeapTypeObject are replaced
by PyTypeObject (an alias thereof).

Instead of the additional private structures, the access is handled
by macros which give runtime access to the structures by computing
the right offset. This way, there is no longer a compile time
dependency of the Python version.

Task-number: PYSIDE-560
Change-Id: Ia2813ba037c4f9c064db5bb10ba10986eb8b473a