Improve libclang detection on Linux
This change essentially allows usage of the distro-specific libclang
libraries for building shiboken. This required changes to the name
of the library that CMake looked for (looking at you Debian).

The change also adds an additional compile test to check for the
availability of the libclang headers which shiboken uses. This should
prevent a failure at build time, and transform it into a nice error
message at cmake configure time.

Change-Id: I37e019fd5923379436a67d77f92c30efd77620d2
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>