Use SPDX name / ID for python license
This is the standard name of the license. See also

  https://spdx.org/licenses/Python-2.0.html

Note that there's some deviation in the actual license text: The
text refers to the exact Python version. Also, the SPDX standardized
license contains additional text for some modules.

Still, it's better to use the standardized name than inventing
our own one.

Change-Id: I64b20bc558928ca28edaf81925a049fcaeb124b6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>