Fix building PySide2 with ninja
When building PySide2 with the ninja generator, ninja prints an error:
 "ninja: error: build.ninja:18549: multiple rules generate
  ./sources/pyside2/pyside_version.py [-w dupbuild=err]"

This seems like an interaction error between CMake and ninja, where
file dependency paths are not canonicalized. In our case
this happens due to usage of pyside_version.py's relative path.

Fix is to use an absolute path.

Task-number: PYSIDE-663
Change-Id: Iee2ae5ea483e59ee67e53e3a2059108bf8a34fbe
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>