Include manylinux1 in package name for limited API builds on Linux
PyPI only accepts binary Linux packages that are built conforming to
PEP 513 / manylinux1 support. The gist of it is that extensions need
to be built on CentOS 5, so that they work on most newer distro
versions than that.

Official Qt can't be built on such an old distro. The minimum
requirement is thus CentOS / RHEL 7 (which we used for packaging
Qt 5.11.0 on Coin).

We do want to upload packages to PyPI, so we have to resort to
including "manylinux1" in the name. Currently this is tied to
the limited API option.

TODO: In the future we should name packages "manylinux1" only
when the distro version is acceptable for Qt build requirements
(RHEL 7.x). It might get a bit messy though, due to
platform.linux_distribution being deprecated, and the necessity
to depend on the new "distro" package.

Change-Id: Ic4dfccd87d810360cbbfce72b27d5fa31e2a59dd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>