Don't pass -DFORCE_LIMITED_API=yes from setup.py unless asked for
It was weird seeing "-DFORCE_LIMITED_API=yes" being passed from
setup.py to CMake when "--limited-api" is not included on the
command line, and then seeing "-DFORCE_LIMITED_API=yes" with Python 2
builds. It gives the impression that maybe it could work with
Python 2.

Instead, don't pass "FORCE_LIMITED_API=yes" from setup.py by default
(but still pass it if it's present on the command line), and instead
default to "yes" within the CMake project themselves.

In the end nothing changes, limited API builds will still be default,
but at least it's not as confusing to see the
"-DFORCE_LIMITED_API=yes" lines for Python 2 builds in the CI log.

Change-Id: I08c863394dd148d88bec324a6cab35459d75303a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>