Adapt examples to proper QByteArray constructor
After correcting the QByteArray behavior, the available
constructors are constraint to bytes, bytearrays,
and others QByteArrays.

The modified lines are methods that use the modified
argument as a QByteArray (interally), that is why we now we
transform them:

"test value" -> b"test value"

Bytes are Python 2 and 3 compatible, so it does not affect
the version.

The only different change is inside `schema.py`
where a change was needed to encode a QByteArray into
utf-8.

Since this is outside our automated test, I manually test
it with Python 2.7.15 and 3.6.3 on Linux.

Task-number: PYSIDE-669
Change-Id: I8c12533b482150f4aee3d813dc127eb6b819549b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>