scriptableapplication: execution as one line
The previous approach executed the entered script line-by-line,
it was not possible to execute multi-line statements,
for example:

for i in range(0, 10):
    print(i)

because PyRun_SimpleString was complaining about the colon.

To avoid all these extra steps we can concatenate all the lines
into one, and then execute it, instead of creating a temporary file.
This will delegate error handling to Python.

Change-Id: Idda572aa1b2e3daad2ba1ed9d70f0a8714b8c995
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>