Fix pyside2-rcc to work on macOS during build process
When preparing the PySide2 package during a build,
regenerate_qt_resources is used on the copied examples to regenerate
the qt resource files for Python3. This uses the pyside2-rcc tool.

The problem was that it failed to find the Qt libraries, due to
missing a rpath entry.

Fix the problem as it is done for shiboken currently, by embedding
the qt rpath value at build time. This get rids of a bunch of error
messages during the build.

Change-Id: I65459bf42818bdac0c6487c6fdd58bf5270d150e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>