Minimize the wheel_override diff
For some shortcomings of bdist_wheel, it was necessary to
write a modified wheel version.

This patch reorders the parts a bit and avoids extra indentation to
make it easier to compare the resulting diff to bdist_wheel.py with
a simple compare tool like sublimerge.

The semantics of the patch is not changed at all.

The patch itself can be best inspected using a diff tool which ignores
whitespace.

-----
We should consider to submit a patch to python.org that has
the necessary improvements for the wheel format, so we can get rid
of this file.

Change-Id: I1f54ad7cb93d64b57462311e5334c906e27d84a0
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>