Create hinting stubs for Python IDEs
This implementation formats all signatures in a way
that is known as type hinting files (.pyi).

Usage
-----

The script is to be called by the same Python interpreter that
was used to build PySide. It works with Python 2 and 3.
On Python 3, it performs a self-test.

    python3  sources/pyside2/PySide2/support/generate_pyi.py run

will generate .pyi files for all compiled PySide modules
and places them into site packages to the binaries.
An optional outpath can be specified.

It is planned to call this script automatically after install.

o Local constants are not included, yet.
  Maybe they never will, unless requested.

o The keyword "from" appears 43 times in argument lists.
  It is fixed in Python, only which does not matter.

o When using Python 3.7 or above, it respects Pep 563 and
  avoids imports which are deferred to runtime.

Task-number: PYSIDE-735
Change-Id: I3bcd5d9284b853fe955376bf35c7897e3698da2b
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>