shiboken: Fix non-deterministic order of some SBK type indexes
Change underlying type of the type database from
a QHash<qualified_name, list of entries> to
a QMultiMap<qualified_name, entry>.

Previously, there was an allEntries() accessor and a function
named entries() building a QHash<qualified_name, entry>.
Simplify this so that there is only an entries() accessor
returning the QMultiMap.

Refactor the various Typedatabase::find() functions to operate on an
iterator range of the QMultiMap.

This unearthed some bugs:

1) In the generators, the call to findType(packageName()) would return
   the namespace entry for "sample" instead of the intended
   module type entry named "sample" due to the ordering.

   Add a new function to search for module type entries and assert
   that it finds it.

2) There was a duplicate, empty primitive type entry for
   QModelIndexList.

Task-number: PYSIDE-757
Change-Id: I1814e4ca67d306e1488398507707cfd07b3f2c78
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>