Enforce pure PyDict to be associated to VariantMap
Classes that implement the dictionary protocol
needs to be treated as PyObject and not plain dictionaries,
then we can properly transform PyDicts to Variant Maps.

The reason behind this is that it seems that the PyDict_Check
is too permisive and allow objects that are not proper
dictionaries.

The default case by any other Sequence-like object will
be a PyObject.

Task-number: PYSIDE-82
Change-Id: I86e8a29e5b8740f69c9c6f315e9d337e599bd333
Reviewed-by: Christian Tismer <tismer@stackless.com>