Fix help display of backport_inspect (python 2.7)
Python 2.7 has no signatures in its inspect module. We therefore add
the missing stuff for the signature package using 'backport_inspect.py'.

The resulting module is a bit unusual because it is assembled
from different modules at runtime. When trying the code

>>> from PySide2.support.signature import inspect
>>> help(inspect)

in Python2, we got only very little information because of some
heuristics in the pydoc module that checks the module identity
and finds only objects from the new one.

It turned out that this heuristics can be circumvented when the
"__all__" property is provided. pydoc then believes everything.

We now have complete help info if the above module is used.
The original inspect module is not changed at all.

Change-Id: I3f24ada3b9ae9c79ec69a6280ddf3ea78735467b
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>