Place templates per module
There is a main template file that hosts most of the
code-templates we use in the typesystems, but there
were still several templates written in each module-typesystem
file.

The templates that are used on different typesystem
are now placed in:

    sources/pyside2/PySide2/templates/core_common.xml

Additionally, the templates used only by a certain module,
are placed next to it, e.g.:

    sources/pyside2/PySide2/templates/widgets_common.xml

Some simple templates were moved to the snippets files too.

Task-number: PYSIDE-799
Change-Id: I4f355cf2ae983aba9e85414e910550189cd18dcb
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>