Move inject-code snippets to separate files
When code must be modified or added to certain
functions, we currently have two options:

    * Write a template and include it into the typesystem,
    * inject code directly at a certain place of the function entry.

The problem with injecting raw code is that since it is an XML
file, one need to escape protected symbols, like "<" and "&".

Thanks to a couple of complementary commits, now we can place
this code snippets on a separate file without triggering
a whole module compiling process, since the typesystem itself
will not be alterered.

All the injected code from the typesystem was removed
and placed into a single file per module inside a top-level directory
called glue.

Some small fixes were included mainly regarding styling
of the code.

Task-number: PYSIDE-834
Change-Id: I823e92d241a528d75d5940090008db9bf297f49e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>