Signature: Implement Nested Classes and Functions for Shiboken
This patch again contains a complete overhaul of the signature
module. The code was re-implemented to properly support nested
classes. Also, the code was reduced by AutoDecRef and by
adopting a concise C++ style.

Note.. We will add a shiboken signature test and complete
mapping.py after the split into three projects is done. The split
changes a lot and is needed right now!

Signatures were quite complete for PySide, but the support for Shiboken
was under-developed.

Since we are planning to generally enhance error messages by using
the Signature module, we should be able to rely on them to always
produce a signature. Therefore, a general overhaul was needed
to resolve all cornes cases for Python 2 and 3.

Nested classes are supported, as well as plain module functions.

The usage of the typing module might improve over time, but the
Signature implementation is now considered complete.

The loader will respect now the path settings which might not be
the package dir but the build dir. This is more consistens with COIN
testing.

Task-number: PYSIDE-795
Change-Id: I246449d4df895dadf2bcb4d997eaa13d78463d9b
Reviewed-by: Simo Fält <simo.falt@qt.io>