Sort writing of type indices when generating module header file
Due to some unknown determinism, sometimes the type indices written
to the generated module header file have a slightly different order,
which means that many cpp files need to be rebuilt for no reason
(semantically the content of the header file does not change).

Make sure to sort the class list by type indices, to try and make
the generated header file as deterministic as possible.

This is a pre-requisite for improved incremental builds.

Change-Id: Ie6a334453cdbfbb601fbac4b6be9291a746650f4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>