Remove createIndex method that uses PyObject*
We have a proper `void*` treatment, so the extra
function does not seems to be necessary.

The injected code used for the interalPointer method was
also removed since is better to leave on shiboken hands
the conversion between void* and python types.

A couple of test were adapted since wrongly used ModelIndex
with Python objects.

Change-Id: I082637c7d1dd1bb0f50a3c453513e5d722f4b783
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>