Improve enum type operations implementation
The current implementation of the enum operations,
wrongly assumes that the first element is always an enum.

This patch add some logic to determinate the types
we are dealing with, to allow operations like:

    2 + QtCore.Qt.Key.Key_1

which were not accepted before.

Float numbers are not accepted for enum operations
and a test case was included.

Some tests were adapted since they were wrongly
implemented.

Fixes: PYSIDE-830
Change-Id: I407dca2b7c39fc684dbdac19ad45d259403ebadd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>